requirements.txt
download
fvcore
lib
matplotlib
numpy
opencv_python
pandas
Pillow
ptflops
pycocotools
pytorch_lightning
rfconv
scikit_image
scikit_learn
seaborn==0.11.2
skimage
tabulate
termcolor
torch
torchmetrics
torchvision
tqdm
wandb
